+++ packages/yarnpkg-core/sources/Configuration.ts
@@ -290,11 +290,6 @@ export const coreDefinitions: {[coreSettingName: string]: SettingsDefinition} = 
     type: SettingsType.BOOLEAN,
     default: true,
   },
-  preferAggregateCacheInfo: {
-    description: `If true, the CLI will only print a one-line report of any cache changes`,
-    type: SettingsType.BOOLEAN,
-    default: true,
-  },
   preferInteractive: {
     description: `If true, the CLI will automatically use the interactive mode when called from a TTY`,
     type: SettingsType.BOOLEAN,
@@ -602,7 +597,6 @@ export interface ConfigurationValueMap {
   enableInlineBuilds: boolean;
   enableProgressBars: boolean;
   enableTimers: boolean;
-  preferAggregateCacheInfo: boolean;
   preferInteractive: boolean;
   preferTruncatedLines: boolean;
   progressBarStyle: string | undefined;